refactor(operator): inline residual_update into schwarz module
SparseGramianUpdater had a single production consumer (FeMultSchwarzSparse)
and lived in its own 58-line module. Moving it next to the type alias that
uses it removes a pointless boundary and makes the multiplicative Schwarz
wiring readable in one file. Tightened pub fn new to pub(crate) since the
struct is constructed only by build_multiplicative_sparse.
